requirements.txt
pandas==2.2
openpyxl==3.1
python-dotenv==0.19.1
pyarrow==17.0.0